+++ src/main.ts
@@ -1,12 +1,12 @@
 import { NestFactory } from '@nestjs/core';
-import serverlessExpress from '@vendia/serverless-express';
-import { Callback, Context, Handler } from 'aws-lambda';
+
+import helmet from 'helmet';
+
 import { AppModule } from './app.module';
-import helmet from 'helmet'
 
-let server: Handler;
+const port = process.env.PORT || 4000;
 
-async function bootstrap(): Promise<Handler> {
+async function bootstrap() {
   const app = await NestFactory.create(AppModule);
 
   app.enableCors({
@@ -15,19 +15,8 @@ async function bootstrap(): Promise<Handler> {
 
   app.use(helmet());
 
-  await app.init();
-
-  const expressApp = app.getHttpAdapter().getInstance();
-
-  return serverlessExpress({ app: expressApp })
+  await app.listen(port);
 }
-
-export const handler = async (
-  event: any,
-  context: Context,
-  callback: Callback
-) => {
-  server = server ?? (await bootstrap());
-  return server(event, context, callback);
-}
-
+bootstrap().then(() => {
+  console.log('App is running on %s port', port);
+});
