+++ src/main_sls.ts
@@ -1,0 +1,33 @@
+import { NestFactory } from '@nestjs/core';
+import serverlessExpress from '@vendia/serverless-express';
+import { Callback, Context, Handler } from 'aws-lambda';
+import { AppModule } from './app.module';
+import helmet from 'helmet'
+
+let server: Handler;
+
+async function bootstrap(): Promise<Handler> {
+  const app = await NestFactory.create(AppModule);
+
+  app.enableCors({
+    origin: (req, callback) => callback(null, true),
+  });
+
+  app.use(helmet());
+
+  await app.init();
+
+  const expressApp = app.getHttpAdapter().getInstance();
+
+  return serverlessExpress({ app: expressApp })
+}
+
+export const handler = async (
+  event: any,
+  context: Context,
+  callback: Callback
+) => {
+  server = server ?? (await bootstrap());
+  return server(event, context, callback);
+}
+
